requirements.txt
playwright
langchain_openai
langchain
langgraph
langsmith
pydantic
python-dotenv
jsonschema
